+++ ResNet/CPU_ONLY/random_pruning_old_cpuOnly.py
