+++ ResNet/CPU_ONLY/utils_old_localBehavior.py
