+++ ResNet/data_prep.py
@@ -1,0 +1,34 @@
+import os
+from PIL import Image
+from torchvision import transforms
+from torch.utils.data import Dataset, DataLoader
+
+class CustomImageDataset(Dataset):
+    def __init__(self, image_dir, transform=None):
+        self.image_dir = image_dir
+        self.transform = transform
+        self.image_paths = [os.path.join(image_dir, img) for img in os.listdir(image_dir) if os.path.isfile(os.path.join(image_dir, img))]
+
+    def __len__(self):
+        return len(self.image_paths)
+
+    def __getitem__(self, idx):
+        img_path = self.image_paths[idx]
+        image = Image.open(img_path).convert("RGB")
+        if self.transform:
+            image = self.transform(image)
+        return image
+
+def prepare_test_dataloader(test_dir, batch_size=32):
+    transform = transforms.Compose([
+        transforms.Resize(256),
+        transforms.CenterCrop(224),
+        transforms.ToTensor(),
+        transforms.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225]),
+    ])
+    test_dataset = CustomImageDataset(test_dir, transform=transform)
+    test_loader = DataLoader(test_dataset, batch_size=batch_size, shuffle=False, num_workers=4)
+    return test_loader
+
+# Prepare DataLoader for the test images
+test_loader = prepare_test_dataloader('imagenet-1k/data/test_images')
